Module Opencv.Algorithm

type t

val clear : t -> unit

Usage: clear __self

Clears the algorithm state

val empty : t -> bool

Usage: empty __self

Returns true if the Algorithm is empty (e.g. in the very beginning or after unsuccessful read

val save : t -> string -> unit

Usage: save __self filename

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

val get_default_name : t -> string

Usage: get_default_name __self

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.