Module Opencv.Clahe

type t = unit Ctypes_static.ptr

val clear : Algorithm.t -> unit

Usage: clear __self

Clears the algorithm state

val empty : Algorithm.t -> bool

Usage: empty __self

Returns true if the Algorithm is empty (e.g. in the very beginning or after unsuccessful read

val save : Algorithm.t -> string -> unit

Usage: save __self filename

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

val get_default_name : Algorithm.t -> string

Usage: get_default_name __self

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

val apply : ?⁠dst:Cvdata.t -> t -> Cvdata.t -> Cvdata.t

Usage: apply ?dst __self src

Equalizes the histogram of a grayscale image using Contrast Limited Adaptive Histogram Equalization.

Parameter: src: Source image of type CV_8UC1 or CV_16UC1.

Parameter: dst: Destination image.

val set_clip_limit : t -> float -> unit

Usage: set_clip_limit __self clip_limit

Sets threshold for contrast limiting.

Parameter: clip_limit: threshold value.

val get_clip_limit : t -> float

Usage: get_clip_limit __self

val set_tiles_grid_size : t -> size2i -> unit

Usage: set_tiles_grid_size __self tile_grid_size

Sets size of grid for histogram equalization. Input image will be divided into equally sized rectangular tiles.

Parameter: tile_grid_size: defines the number of tiles in row and column.

val get_tiles_grid_size : t -> size2i

Usage: get_tiles_grid_size __self

val collect_garbage : t -> unit

Usage: collect_garbage __self