Module Opencv.Generalized_hough

type t = unit Ctypes_static.ptr

val clear : Algorithm.t -> unit

Usage: clear __self

Clears the algorithm state

val empty : Algorithm.t -> bool

Usage: empty __self

Returns true if the Algorithm is empty (e.g. in the very beginning or after unsuccessful read

val save : Algorithm.t -> string -> unit

Usage: save __self filename

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

val get_default_name : Algorithm.t -> string

Usage: get_default_name __self

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

val set_template1 : ?⁠templ_center:point2i -> t -> Cvdata.t -> unit

Usage: set_template1 ?templ_center __self templ

val set_template2 : ?⁠templ_center:point2i -> t -> Cvdata.t -> Cvdata.t -> Cvdata.t -> unit

Usage: set_template2 ?templ_center __self edges dx dy

val detect1 : ?⁠positions:Cvdata.t -> ?⁠votes:Cvdata.t -> t -> Cvdata.t -> Cvdata.t * Cvdata.t

Usage: detect1 ?positions ?votes __self image

val detect2 : ?⁠positions:Cvdata.t -> ?⁠votes:Cvdata.t -> t -> Cvdata.t -> Cvdata.t -> Cvdata.t -> Cvdata.t * Cvdata.t

Usage: detect2 ?positions ?votes __self edges dx dy

val set_canny_low_thresh : t -> int -> unit

Usage: set_canny_low_thresh __self canny_low_thresh

val get_canny_low_thresh : t -> int

Usage: get_canny_low_thresh __self

val set_canny_high_thresh : t -> int -> unit

Usage: set_canny_high_thresh __self canny_high_thresh

val get_canny_high_thresh : t -> int

Usage: get_canny_high_thresh __self

val set_min_dist : t -> float -> unit

Usage: set_min_dist __self min_dist

val get_min_dist : t -> float

Usage: get_min_dist __self

val set_dp : t -> float -> unit

Usage: set_dp __self dp

val get_dp : t -> float

Usage: get_dp __self

val set_max_buffer_size : t -> int -> unit

Usage: set_max_buffer_size __self max_buffer_size

val get_max_buffer_size : t -> int

Usage: get_max_buffer_size __self