Module Opencv__.Mat

type cmat = unit Ctypes_static.ptr

type t = (int, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Genarray.t

val create : unit -> t

create () is a fresh mat.

val clone : t -> t

clone src is a fresh mat containing the same data as src, but with a different underlying array so that the new mat is independent from src.

val cmat_of_bigarray : t -> cmat

val bigarray_of_cmat : cmat -> t

val copy_cmat_bigarray : cmat -> t -> unit